Allow to run opennms from built source in the build environment
Add JICMP, JICMP6 and JRRD to the build image. Allow developer to build the JICMP, JICMP6 and JRRD from source from a
branch to give full control as development dependency against opennms.

Add entrypoint scripts to run opennms in a service stack with Docker compose. Provide an etc-overlay directory to easily
drop-in custom configuration to overwrite the default.

The entrypoint script is just used in the docker-compose file to give flexibility to enter or run the image in
interactive mode.